renamed flat files -- broke two enslaved tables, however.
