+++ src/utils/flatfiles/voyages/voyages_all_card.json
@@ -210,10 +210,6 @@
         "cell_val": {
           "fields": [
             {
-            "var_name": "enslavers__enslaver__principal_alias",
-              "cell_fn": "literal"
-            },
-            {
               "var_name": "enslavers__name_and_role",
               "cell_fn": "literal"
             }
