+++ src/utils/functions/TableAndCardCollectionsOptions.ts
@@ -1,15 +1,15 @@
 import TRANSATLANTIC_TABLE_FLAT from '@/utils/flatfiles/voyages/voyages_transatlantic_table.json';
 import INTRAAMERICAN_TABLE_FLAT from '@/utils/flatfiles/voyages/voyages_intraamerican_table.json';
 import ALLVOYAGES_TABLE_FLAT from '@/utils/flatfiles/voyages/voyages_all_table.json';
-import ENSLAVED_TABLE from '@/utils/flatfiles/enslaved/enslaved_all_table_menu.json';
+import ENSLAVED_TABLE from '@/utils/flatfiles/enslaved/enslaved_all_table.json';
 import AFRICANORIGINS_TABLE from '@/utils/flatfiles/enslaved/enslaved_african_origins_table.json';
-import TEXAS_TABLE from '@/utils/flatfiles/enslaved/enslaved_texas_table_cell_structure.json';
+import TEXAS_TABLE from '@/utils/flatfiles/enslaved/enslaved_texas_table.json';
 import ENSLAVERS_TABLE from '@/utils/flatfiles/enslavers/enslavers_table.json';
 import CARDS_TRANSATLANTIC_COLLECTION from '@/utils/flatfiles/voyages/voyages_transatlantic_card.json';
 import CARDS_INTRAAMERICAN_COLLECTION from '@/utils/flatfiles/voyages/voyages_intraamerican_card.json';
 import CARDS_ALLVOYAGES_COLLECTION from '@/utils/flatfiles/voyages/voyages_all_card.json';
 import CARDS_ENSLAVED_african_origins from '@/utils/flatfiles/enslaved/enslaved_african_origins_card.json';
-import CARDS_ALLENSLAVED from '@/utils/flatfiles/enslaved/enslaved_all_card_menu.json';
+import CARDS_ALLENSLAVED from '@/utils/flatfiles/enslaved/enslaved_all_card.json';
 import CARDS_ENSLAVERS_COLLECTION from '@/utils/flatfiles/enslavers/enslavers_card.json';
 import CARDS_ENSLAVED_TEXAS from '@/utils/flatfiles/enslaved/enslaved_texas_card.json';
 import {
